% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotGrid}
\alias{plotGrid}
\title{plotGrid}
\usage{
plotGrid(plotlist, ncol, nrow)
}
\arguments{
\item{plotlist}{list containing all plots to arrange}
}
\value{
grid of plots
}
\description{
this function arranges grid of multiple plots
}
\details{
This function arranges smoothing plots based on different parameters.
}
\section{Warning}{

You can add sections if you like
}

\examples{

}
