% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotLambda}
\alias{plotLambda}
\title{plotLambda}
\usage{
plotLambda(time, y, z)
}
\arguments{
\item{z}{}
}
\value{
plot
}
\description{
This function plots the smoothed values z.
}
\details{

}
\section{Warning}{

You can add sections if you like
}

\examples{

}
