% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotY}
\alias{plotY}
\title{plotY}
\usage{
plotY(p, pdat)
}
\arguments{
\item{p}{plot}

\item{pdat}{plotdata}
}
\value{

}
\description{
This function returns lambda input
}
\details{

}
\section{}{
NA
}

\examples{

}
