% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothWE.R
\docType{package}
\name{PPGtools}
\alias{PPGtools}
\title{PPGtools}
\description{
This package provides a toolbox for analyzing noisy photoplethysmography (PPG)
raw_signal. The package contains functions for filtering, visualization and feature
extraction from noisy PPG raw_signal. The package contains example PPG raw_signal
(\code{\link{rec}}).
}
\section{Core function}{

The core function can be found at \code{\link{smoothWE}}
}

\author{
Gerbrich Ferdinands (\email{gerbrichferdinands@gmail.com})

Maintainer: Gerbrich Ferdinands (\email{gerbrichferdinands@gmail.com})
}
\keyword{PPG}
