% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInput.R
\name{prepInput}
\alias{prepInput}
\title{prepInput}
\usage{
prepInput(
  raw_data,
  channel = c("Red", "Green", "Blue"),
  tstart = 20,
  tstop = 40
)
}
\arguments{
\item{tstop}{}
}
\value{

}
\description{
.
}
\details{

}
\section{Warning}{

You can add sections if you like
}

\examples{

}
