% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.R
\docType{data}
\name{rec}
\alias{rec}
\title{One example PPG recording}
\format{A data frame with 3200 rows and 7 variables:
\describe{
  \item{time}{timesteps of the recording, in seconds}
  \item{Red}{red signal}
  \item{Green}{red signal}
  \item{Blue}{red signal}
  \item{X}{movement on x-axis}
  \item{Y}{movement on y-axis}
  \item{Z}{movement on z-axis}
}}
\source{
\url{http://www.heartrateapp.com/}
}
\usage{
data(rec)
}
\description{
A dataset containing one example PPG recording of 90 seconds
}
\keyword{datasets}
