% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEsmooth.R
\name{smoothWE}
\alias{smoothWE}
\title{smoothWE}
\usage{
smoothWE(data, lambda, d = 2, t, nuni = FALSE)
}
\arguments{
\item{lambda}{tuning parameter(s), numeric or ?:1 matrix.}

\item{d}{order of differences (d = 1, 2, 3, ...)}

\item{t}{time steps (column vector) (sampling positions)}

\item{y}{noisy series to be smoothed (column vector)}
}
\value{
a smooth series z (fit to y)
}
\description{
This function ..
}
\details{
.
}
\section{Warning}{

You can add sections if you like
}

\examples{

\dontrun{fancySum("A", list(c=51))}

}
\seealso{
\code{\link{Reduce}}
}
