% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddMat.R
\name{ddMat}
\alias{ddMat}
\title{ddMat}
\usage{
ddMat(x, d)
}
\arguments{
\item{x}{time steps}

\item{d}{order of differences}
}
\value{
the matrix D, divided differences. D*Y = divided differences of order d.
}
\description{
matrix D, that constructs divided differences of order d oof y when Dy.
}
\details{

}
\section{Warning}{

You can add sections if you like
}

\examples{

}
