% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBase.R, R/plotLambda.R
\name{plotBase}
\alias{plotBase}
\title{plotBase}
\usage{
plotBase(pdat)

plotBase(pdat)
}
\value{
foundation of a plot

foundation of a plot
}
\description{
This function returns lambda input

This function returns lambda input
}
\details{



}
\section{Warning}{

You can add sections if you like


You can add sections if you like
}

\examples{


}
