% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotBase}
\alias{plotBase}
\title{plotBase}
\usage{
plotBase(pdat)
}
\arguments{
\item{pdat}{a list containing data, and other plot characteristics}
}
\value{
a ggplot object
}
\description{
This function creates a blank canvas for plotting a PPG signal.
}
\details{

}
\examples{

}
