% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotLambda}
\alias{plotLambda}
\title{plotLambda}
\usage{
plotLambda(raw_signal, z, title)
}
\arguments{
\item{title}{}
}
\value{
plot
}
\description{
This function plots the smoothed values z.
}
\details{
todo
}
\section{Warning}{

You can add sections if you like
}

\examples{

}
