% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLambda.R
\name{plotZ}
\alias{plotZ}
\title{plotZ}
\usage{
plotZ(p, pdat)
}
\arguments{
\item{pdat}{}
}
\value{
list of plots
}
\description{
This function plots the smoothed values z.
}
\details{

}
\section{Warning}{

You can add sections if you like
}

\examples{

}
