% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEsmooth.R
\docType{package}
\name{PPGtools}
\alias{PPGtools}
\alias{PPGtools-package}
\title{PPGtools}
\description{
This package provides a toolbox for analyzing noisy photoplethysmography (PPG)
data. The package contains functions for filtering, visualization and feature
extraction from noisy PPG data. The package contains example PPG data
(\code{\link{rec}}).
}
\section{Core function}{

The core function can be found at \code{\link{smoothWE}}
}

\author{
Gerbrich Ferdinands (\email{gerbrichferdinands@gmail.com})

Maintainer: Gerbrich Ferdinands (\email{gerbrichferdinands@gmail.com})
}
\keyword{PPG}
