% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothWE.R
\name{smoothWE}
\alias{smoothWE}
\title{smoothWE}
\usage{
smoothWE(raw_signal, lambda, d = 2, uni = TRUE)
}
\arguments{
\item{raw_signal}{a list (resulting from \code{\link{prepInput}}), containing:
\itemize{
\item Y - a 2 column matrix containing the time steps in the signal ("time") and the noisy series y to be smoothed;
\item channel - Red, Green or Blue;
\item time - the start and end point from the noisy series.
}}

\item{lambda}{tuning parameter(s), numeric or ?:1 matrix.}

\item{d}{order of differences (d = 1, 2, 3, ...)}

\item{uni}{logical indicator of if we equal timesteps are assumed.}
}
\value{
a list containing
\itemize{
\item{z}{- matrix of smooth series z (fit to y)}
\item{lambda}
\item{difference}
\item{uni}
\item{raw_signal}
}
}
\description{
This function ..
}
\examples{


}
